requirements.txt
ollama>=0.4.0
openai>=1.30.0
github-copilot-sdk
python-dotenv>=1.0.0
chromadb>=0.5.0
rich>=13.0.0
pydantic>=2.0.0
